add max_level+change i32 to usize

+++ src/utils/mod.rs
@@ -1,4 +1,4 @@
-pub fn get_printable_coords(nums: &Vec<i32>) -> String {
+pub fn get_printable_coords(nums: &Vec<usize>) -> String {
     nums.into_iter().map(|num| {
         num.to_string()
     }).collect::<Vec<String>>().join(", ").to_string()
